pyproject.toml
[project]
name = "projeto-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcelo Ricardo",email = "marcello.8@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "apache-airflow (>=3.1.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "apache-airflow-providers-postgres (>=6.3.0,<7.0.0)",
    "astro-sdk-python[duckdb] (>=1.8.1,<2.0.0)",
    "airflow-provider-duckdb (>=0.2.0,<0.3.0)",
    "duckdb (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
